requirements.txt
requests==2.23.0
beautifulsoup4==4.9.0
lxml==4.5.2
PyExecJS==1.5.1
